# handlers/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: 心境书签 API
  description: |
    微信小程序「心境书签」后端。

    经网关访问时走 `/bookmark` 前缀。登录接口需要真实的微信 `js_code`，本地可先测「文学推荐」和「健康检查」。
    需要登录的接口：右上角 **Authorize** 填 `Bearer` JWT。
  version: "1.0.0"
servers:
  - url: https://api.soupcircle.xyz/bookmark
    description: 公网（经网关）
  - url: /bookmark
    description: 相对路径 /bookmark
  - url: /
    description: 本地无前缀
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      properties:
        code:
          type: integer
          example: 0
        msg:
          type: string
          example: success
        data: {}
...
paths:
...
  /wxacode:
    get:
      tags: [分享]
      summary: 获取小程序码公网地址
      description: |
        服务端调用 `wxa/getwxacodeunlimit`（page=`pages/index/index`，scene=`share`，width=280）。
        同一 page+scene+env 只生成一次并缓存到 R2。微信失败 `code=1004`。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: code=0 时 data.wxacode_url 如 `/wxacode/share.png`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /wxacode/{filename}:
    get:
      tags: [分享]
      summary: 下载小程序码 PNG
      description: 无需登录，供 canvas 加载。目前仅 `share.png`。
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
            example: share.png
      responses:
        "200":
          description: PNG 图片
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: 尚未生成或不存在
...